+++ src/__tests__/userChangemakerPermissions.int.test.ts
@@ -12,7 +12,7 @@ import {
 	mockJwt as authHeader,
 	mockJwtWithAdminRole as authHeaderWithAdminRole,
 } from '../test/mockJwt';
-import { keycloakUserIdToString, Permission } from '../types';
+import { keycloakIdToString, Permission } from '../types';
 import { NotFoundError } from '../errors';
 
 describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
@@ -22,10 +22,11 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
 				)
 				.send({})
 				.expect(401);
@@ -36,10 +37,11 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
 				)
 				.set(authHeader)
 				.send({})
@@ -58,7 +60,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const user = await loadTestUser();
 			await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/notanId/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/notanId/permissions/${Permission.MANAGE}`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send({})
@@ -69,7 +71,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const user = await loadTestUser();
 			await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/1/permissions/notAPermission`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/1/permissions/notAPermission`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send({})
@@ -81,11 +83,12 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 
 			const response = await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send({})
@@ -104,6 +107,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await createOrUpdateUserChangemakerPermission({
 				userKeycloakUserId: user.keycloakUserId,
@@ -113,7 +117,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			});
 			const response = await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
 				)
 				.set(authHeader)
 				.send({})
@@ -133,6 +137,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await createOrUpdateUserChangemakerPermission({
 				userKeycloakUserId: user.keycloakUserId,
@@ -142,7 +147,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			});
 			const response = await request(app)
 				.put(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
 				)
 				.set(authHeader)
 				.send({})
@@ -163,10 +168,11 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
 				)
 				.send()
 				.expect(401);
@@ -177,10 +183,11 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
 				)
 				.set(authHeader)
 				.send()
@@ -201,7 +208,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const user = await loadTestUser();
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/notanId/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/notanId/permissions/${Permission.MANAGE}`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send()
@@ -212,7 +219,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const user = await loadTestUser();
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/1/permissions/notAPermission`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/1/permissions/notAPermission`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send()
@@ -224,10 +231,11 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.MANAGE}`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send()
@@ -239,6 +247,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await createOrUpdateUserChangemakerPermission({
 				userKeycloakUserId: user.keycloakUserId,
@@ -253,7 +262,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			);
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send()
@@ -265,6 +274,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await createOrUpdateUserChangemakerPermission({
 				userKeycloakUserId: user.keycloakUserId,
@@ -286,7 +296,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			});
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
 				)
 				.set(authHeaderWithAdminRole)
 				.send()
@@ -305,6 +315,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			const changemaker = await createChangemaker({
 				taxId: '11-1111111',
 				name: 'Example Inc.',
+				keycloakOrganizationId: null,
 			});
 			await createOrUpdateUserChangemakerPermission({
 				userKeycloakUserId: user.keycloakUserId,
@@ -332,7 +343,7 @@ describe('/users/changemakers/:changemakerId/permissions/:permission', () => {
 			});
 			await request(app)
 				.delete(
-					`/users/${keycloakUserIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
+					`/users/${keycloakIdToString(user.keycloakUserId)}/changemakers/${changemaker.id}/permissions/${Permission.EDIT}`,
 				)
 				.set(authHeader)
 				.send()