+++ src/database/migrations/0044-alter-funders-add-keycloak_organization_id.sql
@@ -1,0 +1,2 @@
+ALTER TABLE funders
+  ADD COLUMN keycloak_organization_id uuid UNIQUE;