+++ src/test/utils.ts
@@ -1,5 +1,5 @@
 import { createUser, loadUserByKeycloakUserId } from '../database';
-import { stringToKeycloakUserId } from '../types';
+import { stringToKeycloakId } from '../types';
 
 export const isoTimestampPattern =
 	/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{1,6})?(Z|(\+|-)\d{2}:\d{2})$/;
@@ -23,7 +23,7 @@ export const generateNextWithAssertions = (
 	});
 
 export const getTestUserKeycloakUserId = () =>
-	stringToKeycloakUserId('11111111-1111-1111-1111-111111111111'); // This value is not a reference, it's just a static GUID
+	stringToKeycloakId('11111111-1111-1111-1111-111111111111'); // This value is not a reference, it's just a static GUID
 
 export const createTestUser = async () =>
 	createUser({