+++ src/types/index.ts
@@ -14,7 +14,7 @@ export * from './Funder';
 export * from './Id';
 export * from './JsonObject';
 export * from './JsonResultSet';
-export * from './KeycloakUserId';
+export * from './KeycloakId';
 export * from './Language';
 export * from './Opportunity';
 export * from './PaginationParameters';